Moving parquet logic
